change email app

+++ js/apps/email/pages/email-details.cmp.js
@@ -1,5 +1,8 @@
 import { emailService } from "../service/email.service.js"
+
 export default {
+    name:'email-details',
+
     template: `
         <section v-if="email" class="email-datails">
             <div class="details-btn">
@@ -14,29 +17,29 @@ export default {
             <div class="details-body">
                 <p>{{email.body}}</p>
             </div>
-
         </section>
     `, 
+
     data() {
         return {
             email:  null
         }
     }, 
+
     created() {
         const id = this.$route.params.emailId
         emailService.getById(id)
-            .then(email=> {if(email) this.email = email
-                    else emailService.getSentById(id)
-                            .then(email=> this.email = email)
-                })
-                
-        
+            .then(email=> {this.email = email})
     }, 
+    
     methods: {
         removeEmail() {
-            emailService.removeEmail(this.email.id)
-            this.$router.push('/email')
-            // emailService.getById(this.email.id)
+            emailService.getById(this.email.id)
+                .then(email=> {
+                    if(email.status === 'trash') emailService.removeEmail(this.email.id)
+                    else emailService.removeAt(this.email.id)
+                    this.$router.push('/email')
+                })
         }
     }
 }
